CircularList

Description

A CircularList that uses a linked list based on Data.Vector.

Here are a couple of working/design assumptions ...

the inital list is consecutive, means it contains the numbers from 1 to N in a/the given order. And the smallest element is 1.

the linked-list will always have one element it: (0,0). With this element in it the CircularList is considered empty. We need this element because Vector is to the base of 0 and our lists are starting with 1.

to remove an element from the list the predecessor and the successor need to be re-linked. The element it-self will be marked as deleted by making it point to 0.

we will never addinsert an elementitem that is bigger than N.

Documentation

data AlreadyExistsException

Constructors

AlreadyExistsException

data ItemNotFoundException

Constructors

ItemNotFoundException

data ListIsEmptyException

Constructors

ListIsEmptyException

data OutOfRangeException

Constructors

OutOfRangeException

data StackIsEmptyException

Constructors

StackIsEmptyException

type Current = Int

type Item = Int

data CircularList

The list.

Constructors

CircularList Current (Vector Item) [Current]

fromList :: Int -> [Int] -> CircularList

Make a new CircularList (of size N) from a list.

toList :: CircularList -> [Int]

Turn the circular list into a list (from current).

size :: CircularList -> Int

Get the length of the list.

allocated :: CircularList -> Int

Get the allocated (max) length of the list.

isEmpty :: CircularList -> Bool

Check, if list is empty.

isIn :: Int -> CircularList -> Bool

Check, if item is in the list.

get :: CircularList -> Item

Get the current element.

forward :: CircularList -> CircularList

Moving current forward.

insert :: Int -> CircularList -> CircularList

Insert item into list (after current).

Note: The new item cannot be already in the list. Note: Item needs to be in the range (we do not append to the list).

remove :: CircularList -> CircularList

Remove the item from the list (after the current item).

move :: Int -> CircularList -> CircularList

Move current to ...

push :: CircularList -> CircularList

Push the current item on the stack.

pop :: CircularList -> CircularList

Pop the stack and move to that item.