Day05

Description

General - Good one. Simple recurrsion problem. Take a range of N and binary walksearch it (lower, upper) until you findhit the element.

N / L U

A path has the range and the steps in it. A boarding pass has two pathes: A row path and a column path.

Part 1 - Walk the paths for the boarding pathes and find the max of all seat ids.

Part 2 - Find the one seat that is not on the plane.

Documentation

data Step

Constructors

Lower

Upper

data Range

Constructors

Range Int Int

data Path

Constructors

Path Range [Step]

data BoardingPass

Constructors

BoardingPass Path Path

input :: String -> [BoardingPass]

makeSeatId :: BoardingPass -> Int

walkPath :: Path -> Int

part1 :: [BoardingPass] -> Int

part2 :: [BoardingPass] -> Int