Day06

Description

General - Counting problem.

Part 1 - Counting number of questions per group where anybody (any) said yes. Sum up the questions.

Part 2 - Counting number of questions per group where everybody (all) said yes. Sum up the questions.

Documentation

type Group = [String]

input :: String -> [Group]

Read the input file.

part1 :: [Group] -> Int

Solve part1.

part2 :: [Group] -> Int

Solve part2.