Day07

Description

General - Hhhmmm ... to get started, I am going to build two data structures: contains and then (from contains) the reverse of it (isIn).

Part 1 - Just use isIn (recursively) to find all outer bags for a given bag.

Part 2 - Just use contains (recursively) to count how many bag colors can eventually contain at least one shiny gold bag.

Documentation

type Bag = String

type OuterBag = Bag

type InnerBag = [Bag]

type Contains = Map OuterBag [InnerBag]

type IsIn = Map Bag [Bag]

data Bags

Constructors

Bags Contains IsIn

input :: String -> Bags

collectOuters :: IsIn -> Bag -> [Bag]

countingBags :: Contains -> InnerBag -> Int

part1 :: Bags -> Int

part2 :: Bags -> Int