Day12

Description

General - Another "execute instructions" problem. If there is a finite set of instructions you can iterate/fold over the instructions to produce the result. If there are jumps and loops and the end of the program is determined by another condition you probably need to go with a recursion.

Part 1 - This time around we can build a list of instructions. And can just fold the instructions into a/the result. The fold will start with a position of (0, 0) and when it is done we just need to calc the manhatten distance from the final position.

Part 2 - Almost the same as part 1, but ... taking the waypoint logic into consideration.

Documentation

type Argument = Int

type Position = (Int, Int)

data Direction

Constructors

North

South

East

West

type Waypoint = (Int, Int)

type State = (Position, Direction)

type State' = (Position, Waypoint)

data Operation

Constructors

North'

South'

East'

West'

Left'

Right'

Forward'

data Instruction

Constructors

Instruction Operation Argument

makeOp :: Char -> Operation

input :: String -> [Instruction]

manhatten :: Position -> Position -> Int

execute :: State -> Instruction -> State

part1 :: [Instruction] -> Int

execute' :: State' -> Instruction -> State'

part2 :: [Instruction] -> Int