Day18

Description

General - We kinda need to solve two problems here: We need to parse the expressions and we need to evaluate the expressions.

Looking at the expressions you can see an (binary) abstract syntax tree (AST) where every node is an expression. The (two) children are either a value or an expression again.

To build the AST we first convert the expression string from infix notation to postfix notation (RPN) (with the shunting-yard algorithm).

Building the AST from RPN is trivial (Note: We make this easy by reversing the RPN (12+ -> +21)).

To evaluate the AST we just have to (recursively) walk the tree.

Part 1 - Evaluate all expressions and sum up the results.

Part 2 - ???

Documentation

data Expression

Constructors

Add Expression Expression

Mul Expression Expression

Val Int

Nil

pushS :: Char -> String -> String

popS :: String -> (Char, String)

pushQ :: Char -> String -> String

toPostfix :: String -> String

toPostfixToken :: String -> String -> String -> String

toPostfixToken' :: String -> String -> (String, String)

toExpression :: String -> Expression

reverse' :: String -> String

input :: String -> [Expression]

eval :: Expression -> Int

size :: Expression -> Int

part1 :: [Expression] -> Int

part2 :: [Expression] -> Int