Day22

Description

General - Playing cards. With 2 players. In rounds. Recursivly until one of the players has no more cards to play.

Part 1 - Found out who won and calculate the score.

Part 2 - ???

Documentation

type Card = Int

data Game

Constructors

Game [Card] [Card]

input :: String -> Game

Read the input file and return the game to play.

score :: [Int] -> Int

Return/Calculate the score for a given deck.

playGame :: Game -> Game

Play game (by playing rounds recursively)

playRound :: Game -> Game

Play one round.

part1 :: Game -> Int

Solve part1.

playGame2 :: [Game] -> Game -> ([Game], Game)

Play game (by playing rounds recursively) (part2).

playRound2 :: Game -> Game

Play one round (part2).

playSubGame2 :: Game -> Game

Play a subgame (part2).

part2 :: Game -> Int

Solve part2.