Day24

Description

General - To solve this you have to think about (or google) how to navigatewalk a hex-grid (https:trits.ch3bJeNeZ).

To solve this we walk all the steps for any given tile and will end up with a/the list of destination tiles.

Part 1 - Look for all destination tiles that are there/where flipped an odd number of times.

Part 2 - ???

Documentation

type Step = String

type Steps = [Step]

type Tiles = [Steps]

data Position

Constructors

Position Int Int

input :: String -> Tiles

Read the input file and return the paths to the tiles to turn.

walk :: Steps -> Position -> Position

Walk a step (until there are no more steps) and (at the end) return the final position.

blackTiles :: Tiles -> [Position]

Take a list of tile paths, flip them over and return a/the list of black tiles/positions.

part1 :: Tiles -> Int

Solve part1.

part2 :: Tiles -> Int

Solve part2.