Day21

Description

General - Analyse food. Food has incredients. A given allergene is in one-of the ingredients.

Part 1 - Build the intersections to find out which incredients do have allergens. Find more by looking for allergens with only one ingredient (search by elimination; sudoku strategy). You need to do theses two steps/phases recursively until no more "singletons" can be found.

Part 2 - ???

Documentation

type Allergene = String

type Ingredient = String

type Foods = Map Allergene [[Ingredient]]

input :: String -> Foods

intersect :: Eq a => [a] -> [a] -> [a]

intersect' :: Eq a => [[a]] -> [a]

diffr :: Eq a => [a] -> [a] -> [a]

diffMapValues :: Eq v => [v] -> Map k [[v]] -> Map k [[v]]

isSingleton :: forall a. [a] -> Bool

isSingleton' :: forall a. [[a]] -> Bool

flatten :: forall a. [[a]] -> [a]

removeIngredientsByIntersection :: Foods -> Foods

removeIngredientsBySingleton :: Foods -> Foods

part1 :: Foods -> Int

part2 :: Foods -> Int